% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{abundances}
\alias{abundances}
\alias{abundances.mpaqt_quant_result}
\title{Extract raw abundances from quantification result}
\usage{
abundances(x, ...)

\method{abundances}{mpaqt_quant_result}(x, ...)
}
\arguments{
\item{x}{An mpaqt_quant_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Named numeric vector of abundances
}
\description{
Extract raw abundances from quantification result
}
