% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo-bias.R
\name{build_expected_count_matrix}
\alias{build_expected_count_matrix}
\title{Build expected count matrix for bias optimization}
\usage{
build_expected_count_matrix(index, beta, bias_type, quantiles, n_bins, n_ecs)
}
\arguments{
\item{index}{mpaqt_index object}

\item{beta}{Transcript abundances}

\item{bias_type}{"3p" or "5p"}

\item{quantiles}{Distance quantiles}

\item{n_bins}{Number of bins}

\item{n_ecs}{Number of equivalence classes}
}
\value{
Matrix of expected counts per EC and bin
}
\description{
Build expected count matrix for bias optimization
}
\keyword{internal}
