% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-index.R
\name{build_p_list_with_distances}
\alias{build_p_list_with_distances}
\title{Build P list with embedded distance information}
\usage{
build_p_list_with_distances(p_mat, p_trs, p_ecs, distances_dt)
}
\arguments{
\item{p_mat}{Sparse P matrix}

\item{p_trs}{Transcript IDs (column names)}

\item{p_ecs}{EC IDs (row names)}

\item{distances_dt}{Distance information data.table}
}
\value{
List of data.tables, one per transcript
}
\description{
Build P list with embedded distance information
}
\keyword{internal}
