% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-bambu.R, R/validate.R
\name{check_bambu}
\alias{check_bambu}
\title{Check if Bambu is available}
\usage{
check_bambu()

check_bambu()
}
\value{
Invisible TRUE if Bambu is installed

Invisible TRUE if available
}
\description{
Verifies that the bambu Bioconductor package is installed.
}
\keyword{internal}
