% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-post.R
\name{compute_abundance_uncertainty}
\alias{compute_abundance_uncertainty}
\title{Compute Abundance Uncertainty}
\usage{
compute_abundance_uncertainty(index, abundances, fitted_values, sr_counts)
}
\arguments{
\item{index}{mpaqt_index object}

\item{abundances}{Estimated abundances}

\item{fitted_values}{Fitted EC counts}

\item{sr_counts}{Short-read counts}
}
\value{
Named numeric vector of standard errors
}
\description{
Estimate uncertainty in abundance estimates using Fisher information.
}
\keyword{internal}
