% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo-em.R
\name{compute_log_likelihood}
\alias{compute_log_likelihood}
\title{Compute log-likelihood}
\usage{
compute_log_likelihood(
  sr_counts,
  y,
  beta,
  lr_counts,
  lr_probs,
  lr_model,
  prior_var,
  has_lr
)
}
\arguments{
\item{sr_counts}{Short-read EC counts}

\item{y}{Expected EC counts}

\item{beta}{Transcript abundances}

\item{lr_counts}{Long-read counts}

\item{lr_probs}{Long-read coverage probabilities}

\item{lr_model}{Long-read GLM model}

\item{prior_var}{Prior variance}

\item{has_lr}{Whether long-reads are available}
}
\value{
Log-likelihood value
}
\description{
Compute log-likelihood
}
\keyword{internal}
