% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo-bias.R
\name{compute_weighted_expected_counts}
\alias{compute_weighted_expected_counts}
\title{Recompute expected counts with positional weights}
\usage{
compute_weighted_expected_counts(
  index,
  beta,
  weights,
  bias_type,
  quantiles,
  n_ecs
)
}
\arguments{
\item{index}{mpaqt_index object}

\item{beta}{Transcript abundances}

\item{weights}{Positional weights}

\item{bias_type}{"3p" or "5p"}

\item{quantiles}{Distance quantiles}

\item{n_ecs}{Number of equivalence classes}
}
\value{
Expected EC counts
}
\description{
Recompute expected counts with positional weights
}
\keyword{internal}
