% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-counts.R
\name{convert_legacy_counts}
\alias{convert_legacy_counts}
\title{Convert legacy counts format}
\usage{
convert_legacy_counts(counts, type = "auto")
}
\arguments{
\item{counts}{Named numeric vector}

\item{type}{"sr", "lr", or "auto"}
}
\value{
An mpaqt_counts object
}
\description{
Convert legacy counts format
}
\keyword{internal}
