% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-index.R, R/class-index.R
\name{convert_legacy_index}
\alias{convert_legacy_index}
\title{Convert legacy index format}
\usage{
convert_legacy_index(legacy_index)

convert_legacy_index(legacy_index)
}
\arguments{
\item{legacy_index}{List in legacy format}
}
\value{
An mpaqt_index object

An mpaqt_index object
}
\description{
Convert legacy index format

Convert legacy index format to new format
}
\keyword{internal}
