% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{convert_legacy_result}
\alias{convert_legacy_result}
\title{Convert legacy result format}
\usage{
convert_legacy_result(legacy_result)
}
\arguments{
\item{legacy_result}{List in legacy format}
}
\value{
An mpaqt_quant_result object
}
\description{
Convert legacy result format
}
\keyword{internal}
