% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{detect_short_read_input}
\alias{detect_short_read_input}
\title{Detect short-read input type}
\usage{
detect_short_read_input(fastq_1 = NULL, fastq_2 = NULL, rds_file = NULL)
}
\arguments{
\item{fastq_1}{First FASTQ file path}

\item{fastq_2}{Second FASTQ file path}

\item{rds_file}{Pre-processed RDS file path}
}
\value{
Character: "fastq", "fastq_single", "rds", or error
}
\description{
Detect short-read input type
}
\keyword{internal}
