% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-post.R
\name{export_combined_results}
\alias{export_combined_results}
\title{Export Combined Results to CSV}
\usage{
export_combined_results(combined, path)
}
\arguments{
\item{combined}{Combined results from combine_quant_results}

\item{path}{Output file path}
}
\value{
Invisible path
}
\description{
Export Combined Results to CSV
}
