% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{export_results_csv}
\alias{export_results_csv}
\title{Export results to CSV}
\usage{
export_results_csv(
  result,
  path,
  include_abundances = FALSE,
  include_uncertainty = FALSE
)
}
\arguments{
\item{result}{An mpaqt_quant_result object}

\item{path}{Output CSV file path}

\item{include_abundances}{Include raw abundances column}

\item{include_uncertainty}{Include standard errors (if available)}
}
\value{
Invisible path
}
\description{
Export results to CSV
}
