% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-short-read.R
\name{extract_bulk_ec_counts}
\alias{extract_bulk_ec_counts}
\title{Extract Bulk EC Counts from BUS Output}
\usage{
extract_bulk_ec_counts(bus_txt_file, ec_file, index_ec_ids)
}
\arguments{
\item{bus_txt_file}{Path to BUS text file}

\item{ec_file}{Path to matrix.ec file}

\item{index_ec_ids}{EC IDs from the index}
}
\value{
Named numeric vector of EC counts
}
\description{
Extract Bulk EC Counts from BUS Output
}
\keyword{internal}
