% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo-em.R
\name{fit_long_read_model}
\alias{fit_long_read_model}
\title{Fit long-read coverage model}
\usage{
fit_long_read_model(lr_counts, beta, covariates)
}
\arguments{
\item{lr_counts}{Long-read transcript counts}

\item{beta}{Current transcript abundances}

\item{covariates}{Covariate matrix}
}
\value{
List with model and coverage probabilities
}
\description{
Fit Poisson GLM to estimate coverage probabilities from long-read counts.
}
\keyword{internal}
