% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-minimap2.R
\name{get_bam_stats}
\alias{get_bam_stats}
\title{Get BAM Statistics}
\usage{
get_bam_stats(bam_file)
}
\arguments{
\item{bam_file}{Path to BAM file}
}
\value{
List with alignment statistics
}
\description{
Get basic alignment statistics from a BAM file using samtools.
}
\keyword{internal}
