% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{get_input_metadata}
\alias{get_input_metadata}
\title{Get Input Metadata}
\usage{
get_input_metadata(result)
}
\arguments{
\item{result}{An mpaqt_quant_result object}
}
\value{
List with sr_input and lr_input metadata
}
\description{
Get Input Metadata
}
