% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{get_parameters}
\alias{get_parameters}
\title{Get Algorithm Parameters}
\usage{
get_parameters(result)
}
\arguments{
\item{result}{An mpaqt_quant_result object}
}
\value{
List of parameters used in quantification
}
\description{
Get Algorithm Parameters
}
