% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-index.R
\name{get_polya_biotypes}
\alias{get_polya_biotypes}
\title{Get biotypes that have polyA tails}
\usage{
get_polya_biotypes()
}
\value{
Character vector of transcript biotypes
}
\description{
Get biotypes that have polyA tails
}
\keyword{internal}
