% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-system.R
\name{get_tool_version}
\alias{get_tool_version}
\title{Get version of an external tool}
\usage{
get_tool_version(tool)
}
\arguments{
\item{tool}{Name of the tool}
}
\value{
Version string or NULL if cannot determine
}
\description{
Get version of an external tool
}
\keyword{internal}
