% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{has_uncertainty}
\alias{has_uncertainty}
\title{Check if Result Has Uncertainty Estimates}
\usage{
has_uncertainty(result)
}
\arguments{
\item{result}{An mpaqt_quant_result object}
}
\value{
Logical indicating if uncertainty was computed
}
\description{
Check if Result Has Uncertainty Estimates
}
