% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo-bias.R
\name{init_positional_bias}
\alias{init_positional_bias}
\title{Initialize positional bias weights}
\usage{
init_positional_bias(distances, bias_type, n_bins)
}
\arguments{
\item{distances}{Distance data.table from index}

\item{bias_type}{"3p" or "5p" bias type}

\item{n_bins}{Number of distance bins}
}
\value{
List with quantiles and initial weights
}
\description{
Create initial distance bins and weight vector for positional bias correction.
}
\keyword{internal}
