% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-index.R
\name{is_bundled_index}
\alias{is_bundled_index}
\title{Check if Index is Bundled}
\usage{
is_bundled_index(index)
}
\arguments{
\item{index}{mpaqt_index object}
}
\value{
TRUE if index contains embedded Kallisto binary, FALSE otherwise
}
\description{
Check if Index is Bundled
}
\keyword{internal}
