% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-counts.R
\name{length.mpaqt_counts}
\alias{length.mpaqt_counts}
\title{Get number of elements in counts object}
\usage{
\method{length}{mpaqt_counts}(x)
}
\arguments{
\item{x}{An mpaqt_counts object}
}
\value{
Number of elements (ECs for SR, transcripts for LR)
}
\description{
Get number of elements in counts object
}
