% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-index.R
\name{length.mpaqt_index}
\alias{length.mpaqt_index}
\title{Get number of transcripts in index}
\usage{
\method{length}{mpaqt_index}(x)
}
\arguments{
\item{x}{An mpaqt_index object}
}
\value{
Number of transcripts
}
\description{
Get number of transcripts in index
}
