% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-bambu.R
\name{load_bambu_results}
\alias{load_bambu_results}
\title{Load Bambu Results}
\usage{
load_bambu_results(path)
}
\arguments{
\item{path}{Path to saved RDS file}
}
\value{
SummarizedExperiment object
}
\description{
Load saved Bambu SummarizedExperiment from disk.
}
\keyword{internal}
