% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-transcriptome.R
\name{load_bsgenome}
\alias{load_bsgenome}
\title{Load BSgenome Object}
\usage{
load_bsgenome(genome)
}
\arguments{
\item{genome}{BSgenome object or character package name}
}
\value{
BSgenome object
}
\description{
Helper to load a BSgenome object from either a loaded object or package name.
}
\keyword{internal}
