% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-bambu.R
\name{load_genome_for_bambu}
\alias{load_genome_for_bambu}
\title{Load Genome for Bambu}
\usage{
load_genome_for_bambu(genome)
}
\arguments{
\item{genome}{BSgenome object, package name string, or FASTA path}
}
\value{
Genome object suitable for Bambu
}
\description{
Helper to load genome reference for Bambu from various input types.
}
\keyword{internal}
