% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-long-read.R
\name{lr_pathway_bambu_counts}
\alias{lr_pathway_bambu_counts}
\title{Long-Read Pathway: Bambu Count Table}
\usage{
lr_pathway_bambu_counts(bambu_counts, index, sample_id, verbose = TRUE)
}
\arguments{
\item{bambu_counts}{Path to Bambu count table}

\item{index}{mpaqt_index object}

\item{sample_id}{Sample identifier}

\item{verbose}{Print progress messages}
}
\value{
mpaqt_counts_lr object
}
\description{
Long-Read Pathway: Bambu Count Table
}
\keyword{internal}
