% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-long-read.R
\name{lr_pathway_rds}
\alias{lr_pathway_rds}
\title{Long-Read Pathway: Pre-computed RDS}
\usage{
lr_pathway_rds(rds_file, verbose = TRUE)
}
\arguments{
\item{rds_file}{Path to RDS file}

\item{verbose}{Print progress messages}
}
\value{
mpaqt_counts_lr object
}
\description{
Long-Read Pathway: Pre-computed RDS
}
\keyword{internal}
