% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-long-read.R
\name{match_lr_counts_to_index}
\alias{match_lr_counts_to_index}
\title{Match Long-Read Counts to Index Order}
\usage{
match_lr_counts_to_index(counts, index_transcripts)
}
\arguments{
\item{counts}{Named numeric vector of counts}

\item{index_transcripts}{Character vector of transcript IDs in index order}
}
\value{
Named numeric vector in index order
}
\description{
Reorder and fill counts to match index transcript order.
}
\keyword{internal}
