% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-long-read.R
\name{mpaqt_process_long_reads}
\alias{mpaqt_process_long_reads}
\title{Process Long-Read Count File (Bulk)}
\usage{
mpaqt_process_long_reads(count_file, index, output_dir)
}
\arguments{
\item{count_file}{Path to count file}

\item{index}{An \code{mpaqt_index} object}

\item{output_dir}{Output directory}
}
\value{
An \code{mpaqt_counts_lr} object (invisibly)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=mpaqt_prepare_long_reads]{mpaqt_prepare_long_reads()}} instead,
which supports multiple input pathways.
}
\keyword{internal}
