% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-long-read.R
\name{mpaqt_process_long_reads_sc}
\alias{mpaqt_process_long_reads_sc}
\title{Process Long-Read Count File (Single-Cell)}
\usage{
mpaqt_process_long_reads_sc(count_file, clusters_file, index, output_dir)
}
\arguments{
\item{count_file}{Path to count matrix}

\item{clusters_file}{Path to cluster file}

\item{index}{An \code{mpaqt_index} object}

\item{output_dir}{Output directory}
}
\value{
A list of \code{mpaqt_counts_lr} objects (invisibly)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Use \code{\link[=mpaqt_prepare_long_reads_sc]{mpaqt_prepare_long_reads_sc()}} instead.
}
\keyword{internal}
