% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-counts.R
\name{mpaqt_read_counts}
\alias{mpaqt_read_counts}
\title{Load counts from disk}
\usage{
mpaqt_read_counts(path, type = "auto")
}
\arguments{
\item{path}{Path to the RDS file}

\item{type}{Expected type: "sr" (short-read) or "lr" (long-read) or "auto"}
}
\value{
An mpaqt_counts object
}
\description{
Load counts from disk
}
