% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-index.R
\name{mpaqt_read_index}
\alias{mpaqt_read_index}
\title{Load MPAQT Index from Disk}
\usage{
mpaqt_read_index(path)
}
\arguments{
\item{path}{Path to the index RDS file}
}
\value{
An \code{mpaqt_index} object
}
\description{
Read a previously created MPAQT index from an RDS file.
}
\examples{
\dontrun{
idx <- mpaqt_read_index("my_index/mpaqt.index.rds")
print(idx)
}
}
