% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-long-read.R
\name{mpaqt_read_long_read_counts}
\alias{mpaqt_read_long_read_counts}
\title{Load Pre-Processed Long-Read Counts}
\usage{
mpaqt_read_long_read_counts(path)
}
\arguments{
\item{path}{Path to the RDS file}
}
\value{
An \code{mpaqt_counts_lr} object
}
\description{
Load long-read transcript counts from a previously saved RDS file.
}
