% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-pre.R
\name{mpaqt_read_prequant}
\alias{mpaqt_read_prequant}
\title{Load pre-quantification result}
\usage{
mpaqt_read_prequant(path)
}
\arguments{
\item{path}{Path to RDS file}
}
\value{
An mpaqt_prequant object
}
\description{
Load pre-quantification result
}
