% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{mpaqt_read_result}
\alias{mpaqt_read_result}
\title{Load quantification result from disk}
\usage{
mpaqt_read_result(path)
}
\arguments{
\item{path}{Path to the RDS file}
}
\value{
An mpaqt_quant_result object
}
\description{
Load quantification result from disk
}
