% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-short-read.R
\name{mpaqt_read_short_read_counts}
\alias{mpaqt_read_short_read_counts}
\title{Load Pre-Processed Short-Read Counts}
\usage{
mpaqt_read_short_read_counts(path)
}
\arguments{
\item{path}{Path to the RDS file}
}
\value{
An \code{mpaqt_counts_sr} object
}
\description{
Load short-read EC counts from a previously saved RDS file.
}
