% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-pre.R
\name{mpaqt_save_prequant}
\alias{mpaqt_save_prequant}
\title{Save pre-quantification result}
\usage{
mpaqt_save_prequant(prequant, path)
}
\arguments{
\item{prequant}{An mpaqt_prequant object}

\item{path}{Output file path}
}
\value{
Invisible path
}
\description{
Save pre-quantification result
}
