% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-pre.R
\name{new_mpaqt_prequant}
\alias{new_mpaqt_prequant}
\title{Create a new mpaqt_prequant object}
\usage{
new_mpaqt_prequant(
  weights,
  quantiles,
  bias_type,
  n_bins,
  converged,
  iterations,
  initial_beta
)
}
\arguments{
\item{weights}{Positional bias weights}

\item{quantiles}{Distance bin boundaries}

\item{bias_type}{"3p" or "5p"}

\item{n_bins}{Number of bins}

\item{converged}{Whether optimization converged}

\item{iterations}{Number of iterations}

\item{initial_beta}{Initial transcript abundance estimates}
}
\value{
An mpaqt_prequant object
}
\description{
Create a new mpaqt_prequant object
}
\keyword{internal}
