% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-post.R
\name{new_mpaqt_quant_result_enhanced}
\alias{new_mpaqt_quant_result_enhanced}
\title{Create Enhanced mpaqt_quant_result Object}
\usage{
new_mpaqt_quant_result_enhanced(
  abundances,
  tpm,
  log_likelihood,
  converged,
  iterations,
  fitted_values,
  prior_variance,
  positional_weights = NULL,
  positional_bias_type = NULL,
  lr_coverage_probs = NULL,
  lr_model = NULL,
  prior_predictions = NULL,
  normalization = "tpm",
  normalized_counts = NULL,
  uncertainty = NULL,
  parameters = list(),
  sr_input = list(),
  lr_input = NULL
)
}
\arguments{
\item{abundances}{Transcript abundances}

\item{tpm}{TPM values}

\item{log_likelihood}{Log-likelihood}

\item{converged}{Convergence status}

\item{iterations}{Number of iterations}

\item{fitted_values}{Fitted EC counts}

\item{prior_variance}{Prior variance}

\item{positional_weights}{Positional weights}

\item{positional_bias_type}{Bias type}

\item{lr_coverage_probs}{LR coverage probabilities}

\item{lr_model}{LR model object}

\item{prior_predictions}{Prior predictions}

\item{normalization}{Normalization method used}

\item{normalized_counts}{Normalized counts}

\item{uncertainty}{Uncertainty estimates}

\item{parameters}{Algorithm parameters}

\item{sr_input}{Short-read input metadata}

\item{lr_input}{Long-read input metadata}
}
\value{
Enhanced mpaqt_quant_result object
}
\description{
Create Enhanced mpaqt_quant_result Object
}
\keyword{internal}
