% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-post.R
\name{normalize_abundances}
\alias{normalize_abundances}
\title{Normalize Abundances}
\usage{
normalize_abundances(abundances, method = "tpm")
}
\arguments{
\item{abundances}{Raw abundance estimates}

\item{method}{Normalization method}
}
\value{
List with TPM and normalized counts
}
\description{
Normalize Abundances
}
\keyword{internal}
