% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-post.R
\name{normalize_p_matrices_for_umi}
\alias{normalize_p_matrices_for_umi}
\title{Normalize P Matrices for UMI Data}
\usage{
normalize_p_matrices_for_umi(index)
}
\arguments{
\item{index}{mpaqt_index object}
}
\value{
mpaqt_index object with normalized P matrices
}
\description{
Create a working copy of an index with each transcript P matrix normalized
so its probability weights sum to 1.
}
\keyword{internal}
