% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo-bias.R
\name{optimize_positional_weights}
\alias{optimize_positional_weights}
\title{Optimize positional bias weights}
\usage{
optimize_positional_weights(
  sr_counts,
  beta,
  index,
  current_weights,
  bias_type,
  quantiles,
  verbose = FALSE
)
}
\arguments{
\item{sr_counts}{Short-read EC counts}

\item{beta}{Current transcript abundances}

\item{index}{mpaqt_index object}

\item{current_weights}{Current positional weights}

\item{bias_type}{"3p" or "5p"}

\item{quantiles}{Distance quantiles for binning}

\item{verbose}{Print progress}
}
\value{
Updated positional weights
}
\description{
Update positional weights using L-BFGS-B optimization of the Poisson
likelihood.
}
\keyword{internal}
