% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-post.R
\name{prepare_prior}
\alias{prepare_prior}
\title{Prepare Prior for EM Algorithm}
\usage{
prepare_prior(prior_model, index, lr_counts)
}
\arguments{
\item{prior_model}{Prior specification}

\item{index}{mpaqt_index object}

\item{lr_counts}{Long-read counts (optional)}
}
\value{
List with prior predictions
}
\description{
Prepare Prior for EM Algorithm
}
\keyword{internal}
