% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-counts.R
\name{print.mpaqt_counts_lr}
\alias{print.mpaqt_counts_lr}
\title{Print method for mpaqt_counts_lr}
\usage{
\method{print}{mpaqt_counts_lr}(x, ...)
}
\arguments{
\item{x}{An mpaqt_counts_lr object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print method for mpaqt_counts_lr
}
