% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-index.R
\name{print.mpaqt_index}
\alias{print.mpaqt_index}
\title{Print method for mpaqt_index}
\usage{
\method{print}{mpaqt_index}(x, ...)
}
\arguments{
\item{x}{An mpaqt_index object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print method for mpaqt_index
}
