% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-pre.R
\name{print.mpaqt_prequant}
\alias{print.mpaqt_prequant}
\title{Print method for mpaqt_prequant}
\usage{
\method{print}{mpaqt_prequant}(x, ...)
}
\arguments{
\item{x}{An mpaqt_prequant object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print method for mpaqt_prequant
}
