% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{print.mpaqt_quant_result}
\alias{print.mpaqt_quant_result}
\title{Print method for mpaqt_quant_result}
\usage{
\method{print}{mpaqt_quant_result}(x, ...)
}
\arguments{
\item{x}{An mpaqt_quant_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible x
}
\description{
Print method for mpaqt_quant_result
}
