% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-kallisto.R
\name{read_bus_text}
\alias{read_bus_text}
\title{Read BUS text file}
\usage{
read_bus_text(bus_txt_file, columns = c(3, 5))
}
\arguments{
\item{bus_txt_file}{Path to BUS text file}

\item{columns}{Which columns to read (default: ec_id and read_id)}
}
\value{
data.table with requested columns
}
\description{
Parse a BUS file that has been converted to text format.
}
\keyword{internal}
