% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-kallisto.R
\name{read_ec_mapping}
\alias{read_ec_mapping}
\title{Read equivalence class mapping}
\usage{
read_ec_mapping(ec_file)
}
\arguments{
\item{ec_file}{Path to matrix.ec file}
}
\value{
data.table with ec_id and transcript indices
}
\description{
Parse the matrix.ec file produced by kallisto bus.
}
\keyword{internal}
