% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{require_bioc}
\alias{require_bioc}
\title{Require a Bioconductor package}
\usage{
require_bioc(pkg, reason = "for this operation")
}
\arguments{
\item{pkg}{Package name}

\item{reason}{Why it's needed}
}
\value{
Invisible TRUE if available
}
\description{
Require a Bioconductor package
}
\keyword{internal}
