% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-kallisto.R
\name{run_bustools_count}
\alias{run_bustools_count}
\title{Count UMIs from BUS file}
\usage{
run_bustools_count(
  bus_file,
  output_prefix,
  genemap,
  ecmap,
  txnames,
  output_dir = tempdir()
)
}
\arguments{
\item{bus_file}{Path to sorted BUS file}

\item{output_prefix}{Output prefix for count files}

\item{genemap}{Path to transcript-to-gene map}

\item{ecmap}{Path to equivalence class map}

\item{txnames}{Path to transcript names file}

\item{output_dir}{Directory for log files}
}
\value{
Invisible output prefix
}
\description{
Generate count matrices from a sorted, corrected BUS file.
}
\keyword{internal}
