% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-kallisto.R
\name{run_bustools_inspect}
\alias{run_bustools_inspect}
\title{Inspect a BUS file}
\usage{
run_bustools_inspect(bus_file, output_dir = tempdir())
}
\arguments{
\item{bus_file}{Path to BUS file}

\item{output_dir}{Directory for log files}
}
\value{
Invisible bus_file path
}
\description{
Run bustools inspect to get QC statistics from a BUS file.
}
\keyword{internal}
