% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-system.R
\name{run_command_capture}
\alias{run_command_capture}
\title{Run command and capture output}
\usage{
run_command_capture(command, args = character())
}
\arguments{
\item{command}{The system command to execute}

\item{args}{Character vector of arguments}
}
\value{
Character vector of stdout lines
}
\description{
Execute a command and return its stdout as a character vector.
}
\keyword{internal}
