% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-kallisto.R
\name{run_kallisto_index}
\alias{run_kallisto_index}
\title{Run kallisto index}
\usage{
run_kallisto_index(
  transcriptome,
  index_file,
  threads = 1,
  output_dir = tempdir()
)
}
\arguments{
\item{transcriptome}{Path to transcriptome FASTA file}

\item{index_file}{Output path for kallisto index}

\item{threads}{Number of threads to use}

\item{output_dir}{Directory for log files}
}
\value{
Invisible path to created index
}
\description{
Create a kallisto index from a transcriptome FASTA file.
}
\keyword{internal}
