% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra-minimap2.R
\name{run_minimap2_genome}
\alias{run_minimap2_genome}
\title{Run minimap2 for Genome Alignment (for Bambu)}
\usage{
run_minimap2_genome(
  fastq_file,
  genome_fasta,
  output_bam,
  threads = 1L,
  preset = "splice:hq",
  output_dir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{fastq_file}{Path to input FASTQ file}

\item{genome_fasta}{Path to genome FASTA file}

\item{output_bam}{Path for output BAM file}

\item{threads}{Number of threads (default: 1)}

\item{preset}{minimap2 preset (default: "splice:hq")}

\item{output_dir}{Directory for log files}

\item{verbose}{Print progress messages}
}
\value{
Invisible path to output BAM file
}
\description{
Align long-read FASTQ to a genome using minimap2, for use with Bambu.
This produces genome-aligned BAM suitable for transcript discovery/quantification.
}
\keyword{internal}
