% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-pre.R
\name{run_preliminary_em}
\alias{run_preliminary_em}
\title{Run Preliminary EM for Initial Beta Estimates}
\usage{
run_preliminary_em(
  index,
  sr_counts,
  max_iter = 20L,
  tolerance = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{index}{mpaqt_index object}

\item{sr_counts}{Short-read EC counts}

\item{max_iter}{Maximum iterations}

\item{tolerance}{Convergence tolerance}

\item{verbose}{Print progress}
}
\value{
List with beta estimates
}
\description{
Simplified EM algorithm to get initial transcript abundance estimates
for positional weight optimization.
}
\keyword{internal}
