% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-counts.R
\name{save_mpaqt_counts}
\alias{save_mpaqt_counts}
\title{Save counts to disk}
\usage{
save_mpaqt_counts(counts, path)
}
\arguments{
\item{counts}{An mpaqt_counts object}

\item{path}{File path for the RDS file}
}
\value{
Invisible path
}
\description{
Save counts to disk
}
\keyword{internal}
