% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-index.R
\name{save_mpaqt_index}
\alias{save_mpaqt_index}
\title{Save mpaqt_index to disk}
\usage{
save_mpaqt_index(index, path)
}
\arguments{
\item{index}{An mpaqt_index object}

\item{path}{File path for the RDS file}
}
\value{
Invisible path
}
\description{
Save mpaqt_index to disk
}
\keyword{internal}
