% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{save_mpaqt_result}
\alias{save_mpaqt_result}
\title{Save quantification result to disk}
\usage{
save_mpaqt_result(result, path)
}
\arguments{
\item{result}{An mpaqt_quant_result object}

\item{path}{File path for the RDS file}
}
\value{
Invisible path
}
\description{
Save quantification result to disk
}
\keyword{internal}
