% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-short-read.R
\name{sr_pathway_bus}
\alias{sr_pathway_bus}
\title{Short-Read Pathway: BUS + EC Files}
\usage{
sr_pathway_bus(bus_file, ec_file, index, output_dir, verbose = TRUE)
}
\arguments{
\item{bus_file}{Path to BUS file}

\item{ec_file}{Path to EC file}

\item{index}{mpaqt_index object}

\item{output_dir}{Output directory}

\item{verbose}{Print progress messages}
}
\value{
mpaqt_counts_sr object
}
\description{
Short-Read Pathway: BUS + EC Files
}
\keyword{internal}
