% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-short-read.R
\name{sr_pathway_fastq}
\alias{sr_pathway_fastq}
\title{Short-Read Pathway: Raw FASTQ Files}
\usage{
sr_pathway_fastq(fastq_1, fastq_2, index, output_dir, threads, verbose = TRUE)
}
\arguments{
\item{fastq_1}{First FASTQ file}

\item{fastq_2}{Second FASTQ file}

\item{index}{mpaqt_index object}

\item{output_dir}{Output directory}

\item{threads}{Number of threads}

\item{verbose}{Print progress messages}
}
\value{
mpaqt_counts_sr object
}
\description{
Short-Read Pathway: Raw FASTQ Files
}
\keyword{internal}
