% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-short-read.R
\name{sr_pathway_rds}
\alias{sr_pathway_rds}
\title{Short-Read Pathway: Pre-computed RDS}
\usage{
sr_pathway_rds(rds_file, verbose = TRUE)
}
\arguments{
\item{rds_file}{Path to RDS file}

\item{verbose}{Print progress messages}
}
\value{
mpaqt_counts_sr object
}
\description{
Short-Read Pathway: Pre-computed RDS
}
\keyword{internal}
