% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-counts.R
\name{[.mpaqt_counts}
\alias{[.mpaqt_counts}
\title{Extract counts vector}
\usage{
\method{[}{mpaqt_counts}(x, i, ...)
}
\arguments{
\item{x}{An mpaqt_counts object}

\item{i}{Indices or names to extract}

\item{...}{Additional arguments (unused)}
}
\value{
Numeric vector of counts
}
\description{
Extract counts vector
}
