% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-index.R
\name{[.mpaqt_index}
\alias{[.mpaqt_index}
\title{Subset an mpaqt_index by transcript indices}
\usage{
\method{[}{mpaqt_index}(x, i, ...)
}
\arguments{
\item{x}{An mpaqt_index object}

\item{i}{Transcript indices or names}

\item{...}{Additional arguments (unused)}
}
\value{
A subsetted mpaqt_index object
}
\description{
Subset an mpaqt_index by transcript indices
}
