% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-counts.R
\name{summary.mpaqt_counts_sr}
\alias{summary.mpaqt_counts_sr}
\title{Summary method for mpaqt_counts_sr}
\usage{
\method{summary}{mpaqt_counts_sr}(object, ...)
}
\arguments{
\item{object}{An mpaqt_counts_sr object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible summary list
}
\description{
Summary method for mpaqt_counts_sr
}
