% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-index.R
\name{summary.mpaqt_index}
\alias{summary.mpaqt_index}
\title{Summary method for mpaqt_index}
\usage{
\method{summary}{mpaqt_index}(object, ...)
}
\arguments{
\item{object}{An mpaqt_index object}

\item{...}{Additional arguments (unused)}
}
\value{
Invisible summary list
}
\description{
Summary method for mpaqt_index
}
