% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R
\name{tpm}
\alias{tpm}
\alias{tpm.mpaqt_quant_result}
\title{Extract TPM values from quantification result}
\usage{
tpm(x, ...)

\method{tpm}{mpaqt_quant_result}(x, ...)
}
\arguments{
\item{x}{An mpaqt_quant_result object}

\item{...}{Additional arguments (unused)}
}
\value{
Named numeric vector of TPM values
}
\description{
Extract TPM values from quantification result
}
