% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_bias_params}
\alias{validate_bias_params}
\title{Validate positional bias parameters}
\usage{
validate_bias_params(bias_type, bias_bins)
}
\arguments{
\item{bias_type}{Bias type ("3p" or "5p")}

\item{bias_bins}{Number of distance bins}
}
\value{
Invisible TRUE if valid
}
\description{
Validate positional bias parameters
}
\keyword{internal}
