% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_fasta}
\alias{validate_fasta}
\title{Validate a FASTA file}
\usage{
validate_fasta(path)
}
\arguments{
\item{path}{Path to FASTA file}
}
\value{
Invisible TRUE if valid
}
\description{
Validate a FASTA file
}
\keyword{internal}
