% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_fastq}
\alias{validate_fastq}
\title{Validate FASTQ files}
\usage{
validate_fastq(paths, paired = TRUE)
}
\arguments{
\item{paths}{Character vector of FASTQ file paths}

\item{paired}{Expect paired-end files (length 2)}
}
\value{
Invisible TRUE if valid
}
\description{
Validate FASTQ files
}
\keyword{internal}
