% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_gtf}
\alias{validate_gtf}
\title{Validate a GTF file}
\usage{
validate_gtf(path)
}
\arguments{
\item{path}{Path to GTF file}
}
\value{
Invisible TRUE if valid
}
\description{
Validate a GTF file
}
\keyword{internal}
