% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_kallisto_index}
\alias{validate_kallisto_index}
\title{Validate Kallisto Index File}
\usage{
validate_kallisto_index(path)
}
\arguments{
\item{path}{Path to Kallisto index file}
}
\value{
Invisible TRUE if valid
}
\description{
Validate Kallisto Index File
}
\keyword{internal}
