% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_mpaqt_counts}
\alias{validate_mpaqt_counts}
\title{Validate mpaqt_counts object}
\usage{
validate_mpaqt_counts(x, index = NULL)
}
\arguments{
\item{x}{Object to validate}

\item{index}{Optional mpaqt_index to check compatibility}
}
\value{
Invisible TRUE if valid
}
\description{
Validate mpaqt_counts object
}
\keyword{internal}
