% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_mpaqt_counts_lr}
\alias{validate_mpaqt_counts_lr}
\title{Validate long-read counts object}
\usage{
validate_mpaqt_counts_lr(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
Invisible TRUE if valid
}
\description{
Validate long-read counts object
}
\keyword{internal}
