% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_mpaqt_counts_sr}
\alias{validate_mpaqt_counts_sr}
\title{Validate short-read counts object}
\usage{
validate_mpaqt_counts_sr(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
Invisible TRUE if valid
}
\description{
Validate short-read counts object
}
\keyword{internal}
