% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_mpaqt_index}
\alias{validate_mpaqt_index}
\title{Validate an mpaqt_index object}
\usage{
validate_mpaqt_index(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
Invisible TRUE if valid, error otherwise
}
\description{
Validate an mpaqt_index object
}
\keyword{internal}
