% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant-pre.R
\name{validate_mpaqt_prequant}
\alias{validate_mpaqt_prequant}
\title{Validate mpaqt_prequant object}
\usage{
validate_mpaqt_prequant(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
Invisible TRUE if valid
}
\description{
Validate mpaqt_prequant object
}
\keyword{internal}
