% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-result.R, R/validate.R
\name{validate_mpaqt_quant_result}
\alias{validate_mpaqt_quant_result}
\title{Validate mpaqt_quant_result Object}
\usage{
validate_mpaqt_quant_result(x)

validate_mpaqt_quant_result(x)
}
\arguments{
\item{x}{Object to validate}
}
\value{
Invisible TRUE if valid

Invisible TRUE if valid
}
\description{
Validate mpaqt_quant_result Object

Validate quantification result object
}
\keyword{internal}
