% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_normalize}
\alias{validate_normalize}
\title{Validate Normalization Method}
\usage{
validate_normalize(normalize)
}
\arguments{
\item{normalize}{Normalization method}
}
\value{
Invisible TRUE if valid
}
\description{
Validate Normalization Method
}
\keyword{internal}
