% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_output_dir}
\alias{validate_output_dir}
\title{Validate output directory}
\usage{
validate_output_dir(path, create = TRUE)
}
\arguments{
\item{path}{Directory path}

\item{create}{Create if doesn't exist}
}
\value{
Invisible TRUE if valid/created
}
\description{
Validate output directory
}
\keyword{internal}
