% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_technology}
\alias{validate_technology}
\title{Validate technology parameter for single-cell}
\usage{
validate_technology(technology)
}
\arguments{
\item{technology}{Technology string}
}
\value{
Validated technology string
}
\description{
Validate technology parameter for single-cell
}
\keyword{internal}
