% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_umi_correction_timing}
\alias{validate_umi_correction_timing}
\title{Validate UMI Correction Timing}
\usage{
validate_umi_correction_timing(umi_correction_timing)
}
\arguments{
\item{umi_correction_timing}{When to normalize P matrices relative to
positional weighting}
}
\value{
Invisible TRUE if valid
}
\description{
Validate UMI Correction Timing
}
\keyword{internal}
