% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpaqt-package.R
\docType{package}
\name{mpaqt-package}
\alias{mpaqt-package}
\alias{mpaqt}
\title{mpaqt: Multi-Platform Accurate Quantification of Transcripts}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

MPAQT is a generative framework that integrates short- and long-read RNA-seq to overcome each platform's limitations by combining positional-bias modeling with prior integration to enable state-of-the-art isoform-resolved transcript quantification.

RNA-seq quantification pipeline integrating short-read and long-read
sequencing data for improved transcript abundance estimation. Supports
both bulk and single-cell RNA-seq analysis.
}
\section{Main Functions}{

\describe{
\item{\code{\link{mpaqt_index}}}{Create reference index from transcriptome and GTF}
\item{\code{\link{mpaqt_quant_bulk}}}{Quantify bulk RNA-seq samples}
\item{\code{\link{mpaqt_quant_sc}}}{Quantify single-cell RNA-seq samples}
}
}

\section{Algorithm}{

MPAQT uses an EM algorithm to estimate transcript abundances by maximizing:

\deqn{L(\beta) = L_{SR}(\beta) + L_{LR}(\beta) - R(\beta)}

Where:
\itemize{
\item \eqn{L_{SR}}: Short-read likelihood (equivalence class model)
\item \eqn{L_{LR}}: Long-read likelihood (transcript-level, optional)
\item \eqn{R}: L2 regularization penalty
}

Key features include:
\itemize{
\item Positional bias correction (3' or 5' end)
\item Mixed-effects prior integration via gpboost
\item Newton-Raphson optimization for transcript updates
}
}

\section{System Requirements}{

MPAQT requires the following external tools:
\itemize{
\item kallisto (>= 0.50.1) - \url{https://pachterlab.github.io/kallisto/}
\item bustools (>= 0.43.1) - \url{https://bustools.github.io/}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csglab/mpaqt2}
  \item \url{https://csglab.github.io/mpaqt2/}
  \item Report bugs at \url{https://github.com/csglab/mpaqt2/issues}
}

}
\author{
\strong{Maintainer}: Ali Saberi \email{saberiato@gmail.com}

Authors:
\itemize{
  \item Hamed S. Najafabadi [funder]
}

}
\keyword{internal}
